Create an "image_files" directory for Docker
The image_files directory will conform to the layout of files within the Docker
image to improve the ease of understanding where these files end up. As the
`xvfb.init` script was under the old `support` directory, it has been moved into
the new `image_files` directory. The `ADD` line used to copy the file into the
image has also been replaced by a `COPY` line, since the use of `COPY` is
recommended where the special features of `ADD` are not required.