Create a toolchain stage for the jet image
Newer toolchain packages are pulled from the Ubuntu Toolchain Uploads
(restricted) team's test PPA, which is now being added by copying in an
APT sources list file and a GPG public key instead of with
`add-apt-repository`. Duplicated packages have been removed from the
installation list and the `--no-install-recommends` flag is being used
to minimize the installation. Note that in this case, "toolchain" is
being used to refer to any package needed to fetch, build, and install
the project, including all of its third party dependencies.