Create a separate stage for building libserialport
In order to properly segregate the unique build dependencies for
`libserial` from the rest of the `hover-jet` dependencies, its build has
been moved into a separate stage. In this stage, the build dependencies
are installed and the library is cloned and built under `/opt/`. To
install the final products of the build into the main `hover-jet` image,
the built `libserial` repository is temporarily copied over to that
stage, the librarys `install` target is run, and the library is removed.

Preforming builds in isolated stages like this also improves the caching
situation since the ordering of these steps in relation to the rest of
the image's steps no longer invalidate the cache. It also ensures that
no part of the librarys source or temporary files are left over in the
final image, which is something that simply did not happen in the past.

In the future, using Docker's BuildKit will enable us to perform these
independant library builds concurrently, speeding up the build process
for the entire image. Additionally, using the latest experimental
Dockerfile syntax, we will be able to mount the built library from its
stage to run the `install` target, avoiding the needless copying of the
entire repository.