Create a stage for building Eclipse Paho MQTT libs
In order to properly segregate the unique build dependencies for
`paho.mqtt.c` and `paho.mqtt.cpp` from the rest of the `hover-jet`
dependencies, their build has been moved into a separate stage. In this
stage, the build dependencies are installed and the libraries are cloned
and built under `/opt/`. To install the final products of those builds
into the main `hover-jet` image, the built `paho.mqtt.c` and
`paho.mqtt.cpp` repositories are temporarily copied over to that stage,
the libraries `install` targets are run, and the libraries are removed.

Preforming builds in isolated stages like this also improves the caching
situation since the ordering of these steps in relation to the rest of
the image's steps no longer invalidate the cache. It also ensures that
no part of the libraries source or temporary files are left over in the
final image, which is something that simply did not happen in the past.

In the future, using Docker's BuildKit will enable us to perform these
independant library builds concurrently, speeding up the build process
for the entire image. Additionally, using the latest experimental
Dockerfile syntax, we will be able to mount the built libraries from
their stage to run the `install` targets, avoiding the needless copying
of the entire repository.